+++ Sources/App/Models/Scooter.swift
@@ -6,8 +6,8 @@ final class Scooter: SQLiteModel {
     var id: Int?
     var uid: String
 
-    var longitude: Float
-    var latitude: Float
+    var longitude: Double
+    var latitude: Double
     
     var reserved: Bool
     var disabled: Bool
